+++ main.go
@@ -1,4 +1,5 @@
-/*    Copyright 2021 Daniele Bissoli
+/*
+ *   Copyright 2021 Daniele Bissoli
  *
  *   Licensed under the Apache License, Version 2.0 (the "License");
  *   you may not use this file except in compliance with the License.